--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "plugins": ["typecheck", "add-module-exports", "transform-decorators-legacy"],
-  "presets": ["es2015", "react", "stage-0"]
-}
