+++ lib/build.js
@@ -6,6 +6,7 @@
 import {basename, join, relative} from "path";
 import {createGzip} from "zlib";
 import fsExtra from "fs-extra";
+import getBabelBackendConfiguration from "@novivia/babel/config/backend";
 import {Reader as fstreamReader} from "fstream";
 import {getPackageInfo as packageInfo} from "./utils";
 import {isRegExp} from "lodash";
@@ -85,6 +86,7 @@ export class Build {
 
   constructor({
     babelPatterns,
+    compileTargets = {node: 4},
     forceBabelRuntime,
     gzipOptions = {
       level: 6,
@@ -94,6 +96,7 @@ export class Build {
     silent,
     verbose,
   } = {}) {
+    this.compileTargets = compileTargets;
     this.logger = new Logger({
       silent,
       verbose,
@@ -167,27 +170,16 @@ export class Build {
     });
 
     if (isMatch) {
-      const options = {
-        plugins: [
-          require("babel-plugin-add-module-exports"),
-          require("babel-plugin-transform-decorators-legacy").default,
-        ],
-        presets: [
-          require("babel-preset-es2015"),
-          require("babel-preset-react"),
-          require("babel-preset-stage-0"),
-        ],
-      };
-
-      if (this.useBabelRuntime) {
-        options.plugins.push(require("babel-plugin-transform-runtime"));
-      }
+      const babelConfiguration = getBabelBackendConfiguration({
+        targets: this.compileTargets,
+        useRuntime: this.useBabelRuntime,
+      });
 
       this.logger.verbose(`Compiling file: ${relativePath}`);
 
       // This automatically takes care of passing in options found in .babelrc
       // for us, yay!
-      const {code} = await transformBabelFile(relativePath, options);
+      const {code} = await transformBabelFile(relativePath, babelConfiguration);
       await fs.outputFileAsync(
         join(this.temporaryDirectoryPath, relativePath),
         code,
