+++ package.json
@@ -16,7 +16,7 @@
     "url": "git@github.com:Novivia/tools.module-builder.git"
   },
   "scripts": {
-    "build": "./cliWrapper build -b cli.js -p cliWrapper --no-runtime",
+    "build": "./cliWrapper build -b cli.js -p cliWrapper --no-runtime --node-target 7",
     "lint": "novivia-linter lint",
     "pub": "./cliWrapper publish",
     "release": "./cliWrapper release"
@@ -37,16 +37,11 @@
     "@novivia/linter": "^1.6.0"
   },
   "dependencies": {
+    // Novivia dependencies.
+    "@novivia/babel": "0.0.2",
+
+    // npm dependencies.
     "babel-core": "^6.18.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-plugin-typecheck": "^3.9.0",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-register": "^6.18.0",
     "better-log": "^1.3.3",
     "bluebird": "^3.4.6",
     "chalk": "^1.1.3",
@@ -60,6 +55,6 @@
     "tar": "^2.2.1",
     "tempura": "^0.0.8",
     "walkdir": "^0.0.11",
-    "yargs": "^6.3.0"
-  }
+    "yargs": "^6.3.0",
+  },
 }
